create dynamic-layers and move linux-raspberrypi
A variables was added recently to Bitbake (BBFILES_DYNAMIC) which
allows to add bbapends/bb files that are only parsed if a certain
layers is already included.

To be able to share this layer with multiple boards we need to
utilize this mechanism, otherwise we would get errors that typically
say "no matching recipe for bbappend" if we have board/machine
specific bbappends in our layer.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>